chore: CI polish — ruff linting, coverage, shared fixtures (#17)
* refactor: add shared test fixtures in conftest.py

Extract make_mock_engine() and noop_lifespan() into conftest.py,
replacing duplicated code in test_api.py and test_watch.py.

* ci: add ruff linting, pip caching, and coverage reporting

- Add lint job (ruff check + ruff format --check) that gates tests
- Enable pip caching in test jobs
- Add pytest-cov for coverage reporting with artifact upload
- Add pytest-timeout and pytest-cov to dev dependencies
- Configure coverage and 30s default test timeout in pyproject.toml
- Add E402 and E702 to ruff ignore list for script compatibility

* style: auto-fix lint and format across entire codebase

Run ruff check --fix and ruff format on all files to establish
a clean baseline for CI enforcement. Fixes unused imports,
unsorted imports, bare excepts, and formatting inconsistencies.

* fix: move test helpers to helpers.py (conftest.py not importable)

pytest conftest.py files are auto-loaded but cannot be imported
directly. Move shared helpers to tests/helpers.py and add tests/
to pythonpath in pytest config.

---------

+++ src/trio_core/gated_delta.py
@@ -13,9 +13,7 @@ import mlx.nn as nn
 
 @partial(mx.compile, shapeless=True)
 def compute_g(A_log, a, dt_bias):
-    return mx.exp(-mx.exp(A_log.astype(mx.float32)) * nn.softplus(a + dt_bias)).astype(
-        A_log.dtype
-    )
+    return mx.exp(-mx.exp(A_log.astype(mx.float32)) * nn.softplus(a + dt_bias)).astype(A_log.dtype)
 
 
 def _make_gated_delta_kernel(has_mask=False, vectorized=False):
@@ -217,7 +215,12 @@ def gated_delta_ops(
     ys = []
     for t in range(T):
         y, state = _gated_delta_step_ops(
-            q[:, t], k[:, t], v[:, t], g[:, t], beta[:, t], state,
+            q[:, t],
+            k[:, t],
+            v[:, t],
+            g[:, t],
+            beta[:, t],
+            state,
             None if mask is None else mask[:, t],
         )
         ys.append(y)
